frontend/requirements.txt
sentence-transformers
transformers
Flask
elasticsearch
boto3
numpy
 python-dotenv